requirements.txt
discord.py
python-decouple
pymongo[srv]==3.11
motor